requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==1.6.5
attrs==19.1.0
backcall==0.1.0
black==18.9b0
certifi==2018.4.16
Click==7.0
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
django-filter==2.1.0
django-haystack==2.8.1
django-toolbelt==0.0.1
djangorestframework==3.9.1
elasticsearch==6.2.0
gunicorn==19.8.1
ipaddress==1.0.22
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.2.0
pylint==1.9.2
python-dateutil==2.7.3
pytz==2018.4
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
static3==0.7.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
